Create isValidTime.js and IsValidNumber.js, implement isValidNumber in time and date validator scripts, rename checkInput to validateInput in reservation.js
